requirements.txt
fastapi
uvicorn
python-telegram-bot
google-genai
python-dotenv
opik
langgraph
apscheduler
httpx
langchain-google-genai